% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis.R
\name{add_item}
\alias{add_item}
\title{add_item}
\usage{
add_item(object, filename, type, keywords, title, text)
}
\arguments{
\item{object}{The GIS Object that representst the connection to ArcgIS
Online}

\item{filename}{The ZIP file to upload into ArcGIS Online}

\item{type}{Content type, i.e. Shapeile, GDB, etc.}

\item{keywords}{Keywords with which to tag the content}

\item{text}{Text with which to describe the content}
}
\description{
This function is a method to add an item to ArcGIS Online
}
\examples{

add_item(object, file_name, type, keywords, title, text)
Method to add an item to ArcGIS Online

}
\keyword{Add}
\keyword{ArcGIS}
\keyword{Item,}
