% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis.R
\name{add_legend_to_map}
\alias{add_legend_to_map}
\title{add_legend_to_map}
\usage{
add_legend_to_map(m, spdf, colname, colorname = "viridis")
}
\arguments{
\item{m}{The map to add the legend to}

\item{spdf}{The spatial dataframe to be used to drive the legend symbology}

\item{colname}{The column in the features to render as the symbology}

\item{colorname}{The colormap to use for symbology}
}
\description{
This function adds a legend to the map.
}
\examples{

add_legend_to_map(m, spdf, colname, colorname)
This function adds a legend to the map.

}
\keyword{esri,}
\keyword{leaflet,}
\keyword{legend}
\keyword{map,}
