% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis.R
\name{add_polygons_layer}
\alias{add_polygons_layer}
\title{add_polygons_layer}
\usage{
add_polygons_layer(
  spdf,
  colname,
  basemap = providers$Esri.WorldStreetMap,
  colorname = "viridis"
)
}
\arguments{
\item{spdf}{The spatial polygons dataframe to be displayed}

\item{colname}{The column in the features to render as the symbology}

\item{basemap}{An optional basemap parameter can be used to set the basemap}
}
\description{
This function displays a map using geojson polygons as input.
}
\examples{

add_polygons_layer(spdf, colname, basemap)
This function displays a map using geojson polygons as input.

}
\keyword{dataframe}
\keyword{esri,}
\keyword{leaflet,}
\keyword{map,}
\keyword{points,}
\keyword{polygons}
\keyword{spatial}
