% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis.R
\name{add_polylines_layer}
\alias{add_polylines_layer}
\title{add_polylines_layer}
\usage{
add_polylines_layer(spdf, colname, basemap = providers$Esri.WorldStreetMap)
}
\arguments{
\item{spdf}{The spatial polylines dataframe to be displayed}

\item{colname}{The column in the features to render as the symbology}

\item{basemap}{An optional basemap parameter can be used to set the basemap}
}
\description{
This function displays a map using geojson polylines as input.
}
\examples{

add_polylines_layer(spdf, colname, basemap)
This function displays a map using geojson polylines as input.

}
\keyword{dataframe}
\keyword{esri,}
\keyword{leaflet,}
\keyword{map,}
\keyword{points,}
\keyword{polylines}
\keyword{spatial}
