% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis.R
\name{get_column_to_render}
\alias{get_column_to_render}
\title{get_column_to_render}
\usage{
get_column_to_render(attrs)
}
\arguments{
\item{attrs}{The feature attribute dataframe}
}
\description{
This function finds the second numberic column in a feature layers
attributes. It is meant as a kind of 'poor man's smart mapping'
}
\keyword{attributes,}
\keyword{mapping}
\keyword{smart}
