% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis.R
\name{get_first_feature_service}
\alias{get_first_feature_service}
\title{get_first_feature_service}
\usage{
get_first_feature_service(items)
}
\arguments{
\item{items}{The dataframe of items returned by the search funtion}
}
\description{
This finds the first feature service in the dataframe of items returned by
the search function
}
\examples{

get_first_feature_service(items)
This finds the first feature service in the dataframe of items
returned by the search function,
try this get_first_feature_service(items)$url

}
\keyword{ArcGIS,}
\keyword{feature,}
\keyword{hosted,}
\keyword{item,}
\keyword{service}
\keyword{type,}
