% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis.R
\name{get_imagery_sdf}
\alias{get_imagery_sdf}
\title{get_imagery_sdf}
\usage{
get_imagery_sdf(gis, service)
}
\arguments{
\item{service}{The service that you want to query into a dataframe}
}
\description{
This function is a method to get a spatial data frame from an image service
}
\examples{
get_imagery_sdf(service)
}
\keyword{ArcGIS,}
\keyword{dataframe}
