% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis.R
\name{get_item_type}
\alias{get_item_type}
\title{get_item_type}
\usage{
get_item_type(item)
}
\arguments{
\item{item}{The item that you are interested in}
}
\description{
This function is a method get content types and typeKeywords from an item on
AGOL
}
\examples{

get_item_type(item)
Method to get content types and typeKeywords from an item i AGOL

}
\keyword{ArcGIS}
\keyword{item,}
\keyword{type,}
