% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis.R
\name{get_item_types}
\alias{get_item_types}
\title{get_item_types}
\usage{
get_item_types(items)
}
\arguments{
\item{item}{The item list as a dataframe}
}
\description{
This function is a method get a list of AGOL content types and typeKeywords
}
\examples{

get_item_types(items)
Method to get a list of AGOL content types and typeKeywords

}
\keyword{ArcGIS}
\keyword{item,}
\keyword{type,}
