% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis.R
\name{get_sdf}
\alias{get_sdf}
\title{get_sdf}
\usage{
get_sdf(object, service, layer_idx = 0)
}
\arguments{
\item{object}{The connection to your ArcGIS Online account}

\item{service}{The service that you want to query into a dataframe}

\item{layer_idx}{The index of the layer that you are interested in. Defaults
to 0.}
}
\description{
This function is a method to get a spatial data frame from an item
}
\examples{
get_sdf(gis, item)
}
\keyword{ArcGIS,}
\keyword{dataframe}
