% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis.R
\name{get_service}
\alias{get_service}
\title{get_service}
\usage{
get_service(object, item)
}
\arguments{
\item{object}{The connection to your ArcGIS Online account}

\item{item}{The ArcGIS Online item that you are interested in}
}
\description{
This function is  a method to get the service properties
}
\examples{
get_service(gis, item)
}
\keyword{ArcGIS,}
\keyword{feature}
\keyword{service}
\keyword{service,}
