% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis-package.R, R/arcgis.R
\docType{class}
\name{GIS-class}
\alias{GIS-class}
\alias{GIS}
\title{GIS class}
\arguments{
\item{url}{The url to your ArcGIS Online organization}

\item{username}{Your ArcGIS Online username}

\item{password}{Your ArcGIS Online password}
}
\description{
This clas contains your ArcGIS Online credentials and session token

This clas contains your ArcGIS Online credentials and session token
}
\examples{

list(url, username, password)

list(url, username, password)
}
\keyword{ArcGIS}
\keyword{GIS,}
