% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis.R
\name{is_hosted_feature_service}
\alias{is_hosted_feature_service}
\title{is_hosted_feature_service}
\usage{
is_hosted_feature_service(item)
}
\arguments{
\item{item}{The item that you are interested in}
}
\description{
This function is determines whether the item you are interested in
}
\details{
is a hosted feature service
}
\examples{
is_hosted_feature_service(item)
Method to determine whether the item you are interested in
is a hosted feature service
}
\keyword{ArcGIS,}
\keyword{feature,}
\keyword{hosted,}
\keyword{item,}
\keyword{service}
\keyword{type,}
