% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis.R
\name{list_esri_basemaps}
\alias{list_esri_basemaps}
\title{list_esri_basemaps}
\usage{
list_esri_basemaps()
}
\description{
This function lists the available esri basemaps.
}
\examples{
list_esri_basemaps()
This function lists the available esri basemaps.
}
\keyword{basemaps}
\keyword{esri,}
\keyword{leaflet,}
\keyword{map,}
