% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis.R
\name{list_stamen_basemaps}
\alias{list_stamen_basemaps}
\title{list_stamen_basemaps}
\usage{
list_stamen_basemaps()
}
\description{
This function lists the available stamen basemaps.
}
\examples{
list_stamen_basemaps()
This function lists the available stamen basemaps.
}
\keyword{basemaps}
\keyword{leaflet,}
\keyword{map,}
\keyword{stamen,}
