% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis.R
\name{login}
\alias{login}
\title{GIS class}
\usage{
login(object)
}
\arguments{
\item{gis}{The connection to your ArcGIS Online account}
}
\description{
This function logs you into ArcGIS ONline
}
\examples{
login(gis)
}
\keyword{ArcGIS,}
\keyword{GIS,}
\keyword{login}
