% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis.R
\name{search_gis}
\alias{search_gis}
\title{search_gis}
\usage{
search_gis(object, q)
}
\arguments{
\item{object}{The connection to your ArcGIS Online account}

\item{q}{The query you are submitting to ArcGIS Online}
}
\description{
This function is a nethod to search ArcGIS Online
}
\examples{
search_gis(gis, query)
}
\keyword{ArcGIS,}
\keyword{query,}
\keyword{search}
