% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis.R
\name{show_map}
\alias{show_map}
\title{show_map}
\usage{
show_map(basemap = providers$Esri.WorldStreetMap)
}
\arguments{
\item{basemap}{An optional basemap parameter can be used to set the basemap}
}
\description{
This function displays a leaflet map. In the future I hope to extent this
function to do more
}
\examples{
show_map(basemap)
This function displays a leaflet map. In the future I hope to extent this
function to do more
}
\keyword{esri}
\keyword{leaflet,}
\keyword{map,}
